+++ lib/osm.dart
@@ -1,6 +1,7 @@
 /// Reading and processing OpenStreetMap data.
 library;
 
+export 'src/area.dart';
 export 'src/bounds.dart';
 export 'src/element.dart';
 export 'src/filter.dart';